pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "churro-ocr"
version = "0.1.0"
description = "Library-first OCR and layout detection for historical documents"
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
authors = [{ name = "History Genie" }]
keywords = ["ocr", "layout-detection", "historical-documents", "llm"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
]
dependencies = [
    "loguru>=0.7.2,<1",
    "Pillow>=10.4.0,<12",
    "rich>=13.9.2,<14",
    "typer>=0.12.3,<1",
]

[project.optional-dependencies]
llm = [
    "google-auth>=2.41.1,<3",
    "litellm[caching]==1.82.3",
]
azure = [
    "azure-ai-documentintelligence==1.0.2",
]
huggingface = [
    "qwen-vl-utils",
    "transformers[torch]>=4.57.0,<5",
    "torchvision",
]
hf = [
    "qwen-vl-utils",
    "transformers[torch]>=4.57.0,<5",
    "torchvision",
]
vllm = [
    "transformers>=4.57.0,<5",
    "torchvision",
    "vllm>=0.18,<1",
]
mistral = [
    "mistralai>=1.6.0,<2",
]
local = [
    "litellm[caching]==1.82.3",
]
pdf = [
    "pypdfium2>=5,<6",
]
all = [
    "google-auth>=2.41.1,<3",
    "litellm[caching]==1.82.3",
    "azure-ai-documentintelligence==1.0.2",
    "qwen-vl-utils",
    "transformers[torch]>=4.57.0,<5",
    "torchvision",
    "vllm>=0.18,<1",
    "mistralai>=1.6.0,<2",
    "pypdfium2>=5,<6",
]

[dependency-groups]
dev = [
    "build",
    "coverage>=7.11.0,<8",
    "datasets>=4.4.1,<5",
    "evaluate>=0.4.6,<0.5",
    "google-auth>=2.41.1,<3",
    "nltk>=3.9.2,<4",
    "pyarabic>=0.6.15,<0.7",
    "pytest",
    "pytest-asyncio",
    "pytest-cov>=7.0.0,<8",
    "rapidfuzz>=3.14.3,<4",
    "ruff",
    "tqdm>=4.67.1,<5",
    "twine",
    "ty>=0.0.24,<0.0.25",
]
minimal = [
    "qwen-vl-utils",
    "transformers[torch]>=4.57.0,<5",
    "torchvision",
]

[project.scripts]
churro-ocr = "churro_ocr.cli:main"

[project.urls]
Homepage = "https://github.com/stanford-oval/Churro"
Repository = "https://github.com/stanford-oval/Churro"
Issues = "https://github.com/stanford-oval/Churro/issues"

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["churro_ocr*"]

[tool.setuptools.package-data]
"churro_ocr" = ["py.typed"]

[tool.pixi.workspace]
channels = ["conda-forge", "main", "r", "msys2"]
name = "churro-ocr"
platforms = ["linux-64"]

[tool.pixi.dependencies]
pip = "*"

[tool.pixi.feature.repo-default.pypi-dependencies]
"churro-ocr" = { path = ".", editable = true, extras = ["llm", "azure", "hf", "mistral", "pdf"] }

[tool.pixi.feature.repo-minimal.pypi-dependencies]
"churro-ocr" = { path = ".", editable = true, extras = ["hf"] }

[tool.pixi.feature.cuda.dependencies]
cudnn = "9.*"
cuda-libraries = "12.*"

[tool.pixi.environments]
default = { features = ["repo-default", "dev"], solve-group = "default" }
minimal = { features = ["dev", "repo-minimal"], solve-group = "default" }
cuda = { features = ["repo-default", "cuda", "dev"], solve-group = "default" }

[tool.pixi.tasks]
format = "ruff format src tests"
lint = "ruff check src tests"
typecheck = "ty check src tests"
test = "python -m pytest"
coverage = "python -m pytest --cov=src/churro_ocr --cov=tooling --cov-report=term-missing"
package-check = "python scripts/package_check.py"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: live provider integration tests that require network and credentials",
]

[tool.ruff]
target-version = "py312"
line-length = 110

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC", "SIM"]

[tool.ruff.lint.per-file-ignores]
"src/churro_ocr/cli.py" = ["B008"]
